Add option to extend a list to an entity list in addition to just appending
When combined with the fact that extending Blueprint.entities using += notation causes a deep copy, there wasn't a clean way to add a list of entities to a blueprint.
This adds one, which testing shows is more performant than the += method, and produces identical results to simply appending in sequence.