Add performance note regarding Blueprint.entities
If you do `blueprint.entities += [foo, bar, baz]` then this triggers a deep copy of the resulting list when it assigns it back to blueprint.entities.
By changing the code to:
```
for entity in [foo, bar, baz]:
    blueprint.entities.append(entity)
```
or just `blueprint.entities.extend([foo, bar, baz])` the performance improves by an order of magnitude. (>1min to <8 seconds on the script I'm developing).

This is (to me) sufficiently unintuitive that it's worth documenting.